+++ src/OpenMRS_Project/Nuri/Nuri_POM.java
@@ -1,19 +1,11 @@
-package OpenMRS_Project;
+package OpenMRS_Project.Nuri;
 
 import Utility.BaseDriver;
 import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Assert;
-
-import java.awt.*;
-import java.awt.event.KeyEvent;
-import java.time.Duration;
-import java.util.List;
 
 import static Utility.BaseDriver.driver;
 import static Utility.BaseDriver.wait;
@@ -97,11 +89,11 @@ public class Nuri_POM {
 
     public void login() {
         driver.navigate().to("https://openmrs.org/demo/");
+        JavascriptExecutor js = (JavascriptExecutor) driver;
 
         demoButton.click();
-        JavascriptExecutor js = (JavascriptExecutor) driver;
         js.executeScript("arguments[0].scrollIntoView(true);", exploreButton);
-        exploreButton.click();
+        js.executeScript("arguments[0].click();",exploreButton);
         wait.until(ExpectedConditions.elementToBeClickable(enterMrsButton));
         js.executeScript("arguments[0].scrollIntoView(true);", enterMrsButton);
         js.executeScript("arguments[0].click();", enterMrsButton);
