+++ src/Utility/BaseDriver.java
@@ -42,8 +42,8 @@ public class BaseDriver {
 
         driver = new ChromeDriver();
         driver.manage().window().maximize();
-        driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(20));
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(20));
+        driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(7));
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(7));
         wait = new WebDriverWait(driver, Duration.ofSeconds(20));
